refactor(desktop): organize tauri shell modules
- Move deep-link, account, OAuth, app-menu, and window helpers into their owning modules.

- Rename desktop support modules to clearer names and update call sites.

- Keep Linux deep-link handler behavior unchanged while reducing lib.rs scope.
